Update debug output

+++ src/Services/Authentication/RemoteAddressAuthenticator.php
@@ -103,8 +103,12 @@ class RemoteAddressAuthenticator
      */
 	public function resetAuth(): string
 	{
-		$this->storageRepository->deleteObject(Settings::PLUGIN_NAME,self::AUTH_FILENAME);
-		return 'DELETED';
+        if($this->storageRepository->doesObjectExist(Settings::PLUGIN_NAME,self::AUTH_FILENAME)) {
+            $this->storageRepository->deleteObject(Settings::PLUGIN_NAME, self::AUTH_FILENAME);
+            return 'DELETED';
+        }
+
+        throw new Exception('No saved whitelist!');
 	}
 
 	/**
